+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - move the relation between project and surveys to be saved in the surveys table instead through the actions table
  - surveys now are synchronized at `surveys` not at `actions`
  - add field `extra_params` to style `qualtricsSurveys`
+ - add `publish` survey functionality
 
 # v1.2.0
 ### Bugfix